fix calculation, end of line

+++ f2m.py
@@ -4,10 +4,10 @@ from tkinter import ttk
 def calculate(*args):
     try:
         value = float(feet.get())
-        meters.set((0.3048 * value * 10000.0 + 0.5)/10000.0)
+        meters.set(int(0.3048 * value * 10000.0 + 0.5)/10000.0)
     except ValueError:
         pass
-    
+
 root = Tk()
 root.title("Feet to Meters")
